split base classes
